Basically done with an initial draft..
Still need to find a way to make it easier to present. I fear I'll need to convert to RST to be able to convert it all to one big PDF.
There are still a lot of links that need to be fixed (TODOs) and background explanation that is missing.
And I'm sure it could use a good proofreading by someone else